+++ tests/test_ssh2_client.py
@@ -4,11 +4,12 @@ import logging
 import time
 import subprocess
 
-from gevent import socket
+from gevent import socket, sleep
 
 from .base_ssh2_test import SSH2TestCase
 from .embedded_server.openssh import OpenSSHServer
 from pssh.ssh2_client import SSHClient, logger as ssh_logger
+from pssh.tunnel import Tunnel
 from ssh2.session import Session
 from pssh.exceptions import AuthenticationException, ConnectionErrorException, \
     SessionError
@@ -56,7 +57,7 @@ class SSH2ClientTest(SSH2TestCase):
         del client.session
         del client.sock
         client.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        client._connect()
+        client._connect(self.host, self.port)
         client._init()
         # Identity auth
         client.pkey = None
@@ -64,7 +65,7 @@ class SSH2ClientTest(SSH2TestCase):
         del client.session
         del client.sock
         client.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        client._connect()
+        client._connect(self.host, self.port)
         client.session = Session()
         client.session.handshake(client.sock)
         self.assertRaises(AuthenticationException, client.auth)
@@ -115,3 +116,23 @@ class SSH2ClientTest(SSH2TestCase):
         self.assertRaises(ConnectionErrorException,
                           SSHClient, self.host, port=12345,
                           num_retries=2)
+
+    ## OpenSSHServer needs to run in its own thread for this test to work
+    ##  Race conditions otherwise.
+    #
+    # def test_direct_tcpip(self):
+    #     proxy_host = '127.0.0.9'
+    #     server = OpenSSHServer(listen_ip=proxy_host, port=self.port)
+    #     server.start_server()
+    #     t = Tunnel(self.host, proxy_host, self.port,
+    #                port=self.port,
+    #                pkey=self.user_key,
+    #                num_retries=1,
+    #                timeout=5)
+    #     t.daemon = True
+    #     t.start()
+    #     while not t.tunnel_open.is_set():
+    #         sleep(.1)
+    #     client = SSHClient('127.0.0.1', port=t.listen_port,
+    #                        pkey=self.user_key,
+    #                        timeout=2)